Removed the "repair.py" file with the (not entirely correct) updated repair algorithm and replaced it with the almost-original "repair.py" file, updated for shapely 2 and with a few minor edits so that some of the  functions will work correctly for non-simply-connected polygons.  This repository is now intended to house the files for a maintenance update for compatibility with shapely 2, and I think it's now complete.  I'll create a separate repository (soon!) for the major update that will feature the new repair algorithm.
Removed the "repair.py" file with the (not entirely correct) updated repair algorithm and replaced it with the almost-original "repair.py" file, updated for shapely 2 and with a few minor edits so that some of the  functions will work correctly for non-simply-connected polygons.

This repository is now intended to house the files for a maintenance update for compatibility with shapely 2, and I think it's now complete.  I'll create a separate repository (soon!) for the major update that will feature the new repair algorithm.